refactor: 카테고리 탭 컴포넌트의 카테고리 목록을 상수로 이동하고 관련 훅을 새로 작성
