+++ src/constants/home.constants.ts
@@ -11,3 +11,12 @@ export const CATEGORY_ROUTES: Record<TravelCategory, string> = {
   카페: '/',
   항공: '/',
 };
+
+export const CATEGORIES: TravelCategory[] = [
+  '전체',
+  '명소',
+  '숙박',
+  '맛집',
+  '카페',
+  '항공',
+];
